Spread Spectrum Engine
A correlator is designed to find the phase and frequency of the input
signal. The input has pseudo random noise added to it. the design unit
first generates a sine wave , then adds pseudo random noise to it. this
signal is then compared to the input signal and their correlation is
obtained by multiply and add method. Thus the match determines the phase
and frequency of the input signal. all the registers are 32 bit each.